refactor: make its child folder structure more clear on `commands` and `events`
